Use parallel download manager and netcdf aggregator to perform aggregations
