+++ src/extension/wps/src/main/java/au/org/emii/download/DownloadRequest.java
@@ -1,0 +1,25 @@
+package au.org.emii.download;
+
+import java.net.URL;
+
+/**
+ * Requested download metadata
+ */
+public class DownloadRequest {
+    private final URL url;
+    private final long size;
+
+    public DownloadRequest(URL url, long size) {
+        this.url = url;
+        this.size = size;
+    }
+
+    public URL getUrl() {
+        return url;
+    }
+
+    public long getSize() {
+        return size;
+    }
+
+}
