--- src/extension/wps/src/main/java/au/org/emii/gogoduck/download/DownloadRequest.java
@@ -1,25 +1,0 @@
-package au.org.emii.gogoduck.download;
-
-import java.net.URL;
-
-/**
- * Requested download metadata
- */
-public class DownloadRequest {
-    private final URL url;
-    private final long size;
-
-    public DownloadRequest(URL url, long size) {
-        this.url = url;
-        this.size = size;
-    }
-
-    public URL getUrl() {
-        return url;
-    }
-
-    public long getSize() {
-        return size;
-    }
-
-}
