--- src/extension/wps/src/main/java/au/org/emii/gogoduck/exception/GoGoDuckException.java
@@ -1,11 +1,0 @@
-package au.org.emii.gogoduck.exception;
-
-public class GoGoDuckException extends RuntimeException {
-    public GoGoDuckException(String reason) {
-        super(reason);
-    }
-
-    public GoGoDuckException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
