--- src/extension/wps/src/main/java/au/org/emii/gogoduck/worker/IndexReader.java
@@ -1,7 +1,0 @@
-package au.org.emii.gogoduck.worker;
-
-import au.org.emii.gogoduck.exception.GoGoDuckException;
-
-public interface IndexReader {
-    URIList getUriList(String profile, String timeField, String sizeField, String urlField, GoGoDuckSubsetParameters subset) throws GoGoDuckException;
-}
