--- src/extension/wps/src/main/java/au/org/emii/gogoduck/worker/Subset.java
@@ -1,11 +1,0 @@
-package au.org.emii.gogoduck.worker;
-
-public class Subset {
-    public String start;
-    public String end;
-
-    public Subset(String start, String end) {
-        this.start = start;
-        this.end = end;
-    }
-}
